Core/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
grpcio==1.59.2
    # via
    #   -r requirements.in
    #   grpcio-tools
grpcio-tools==1.59.2
    # via -r requirements.in
protobuf==4.25.0
    # via grpcio-tools
redis==5.0.1
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
